+++ CurrencyConverter/ObjectModels/Currency.swift
@@ -11,14 +11,14 @@ struct Currency {
     var buy: Double
     var sell: Double
     var code: String
-    var number: Int16
+    var number: Int
     var country: String
-    var groupKey: Int16
+    var groupKey: Int
     var currency: String
     var container: NSOrderedSet?
     var currencyPlural: String
     
-    init(buy: Double, sell: Double, code: String, number: Int16, country: String, groupKey: Int16, currency: String, container: NSOrderedSet? = nil, currencyPlural: String) {
+    init(buy: Double, sell: Double, code: String, number: Int, country: String, groupKey: Int, currency: String, container: NSOrderedSet? = nil, currencyPlural: String) {
         self.buy = buy
         self.sell = sell
         self.code = code
@@ -34,7 +34,7 @@ struct Currency {
         self.buy = 0.0
         self.sell = 0.0
         self.code = dictionary["Code"] as? String ?? String()
-        self.number = dictionary["Number"] as? Int16 ?? Int16()
+        self.number = dictionary["Number"] as? Int ?? Int()
         self.country = dictionary["Country"] as? String ?? String()
         self.groupKey = 0
         self.currency = dictionary["Currency"] as? String ?? String()
